requirements.txt
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
pycryptodome==3.19.0
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.1.0
